+++ webclient/src/components/style/scroll.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const Scroll = styled.div`
+  overflow-x: auto;
+`;