+++ webclient/src/components/style/table.js
@@ -1,0 +1,27 @@
+import styled from "styled-components";
+
+export const Table = styled.table`
+  border-collapse: collapse;
+  width: 100%;
+
+  td {
+    border: 1px solid #ddd;
+    padding: 8px;
+  }
+
+  tr:nth-child(even) {
+    background-color: #f2f2f2;
+  }
+
+  tr:hover {background-color: #ddd;}
+
+  th {
+    border: 1px solid #ddd;
+    padding: 8px;
+    padding-top: 12px;
+    padding-bottom: 12px;
+    text-align: left;
+    background-color: #4caf50;
+    color: white;
+  }
+`;