+++ ChangeLog.md
@@ -1,4 +1,5 @@
-
+### 2019-02-16
+* implemented react-table component with server-side data
 
 ### 2019-02-09
 * moved formHook functions to common location